+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Features
+
+-   formats the contents of jinja tags (the code between the curlies) using *black*, the Python formatter. If *black* is not installed, re-install sqlfmt with the jinjafmt extra (`pipx install sqlfmt[jinjafmt]`). If *black* is installed, but you do not want to use this feature, you can disable it with the command-line option `--no-jinja-fmt`
+
 ## [0.4.2] - 2022-01-26
 
 ### Features